Structured the changes more cleanly
