--- lib/features/profile/lib/profile_screen.dart
@@ -1,27 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'profile_page.dart';
-
-
-class ProfileScreen extends StatelessWidget {
-  const ProfileScreen({
-    super.key,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return MaterialApp(
-      title: "Profile Demo",
-      theme: ThemeData(
-        useMaterial3: true,
-        textTheme: TextTheme(
-          bodyLarge: TextStyle(fontFamily: 'Roboto', fontSize: 18),
-          bodySmall: TextStyle(fontFamily: 'Roboto', fontSize: 14),
-          titleLarge: TextStyle(fontFamily: 'Roboto', fontWeight: FontWeight.bold),
-        ),
-        colorScheme: ColorScheme.fromSeed(seedColor: Colors.green),
-      ),
-      home: ProfilePage(),
-    );
-  }
-}
